requirements.txt
APScheduler==3.6.3
blinker==1.9.0
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.11.3
pyTelegramBotAPI==4.26.0
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.3
setuptools==78.1.0
six==1.17.0
tornado==6.1
tzlocal==5.3.1
urllib3==1.26.15
Werkzeug==3.1.3
